+++ scripts/preprocess.py
@@ -12,27 +12,28 @@ for filename in os.listdir("../data/ipgzip"):
         z = zipfile.ZipFile(fh)
         
         for name in z.namelist():
-            outpath = "../data/ipgxml"
-            z.extract(name, outpath)
+            if not os.path.isfile("../data/ipgxml/" + name):
+                outpath = "../data/ipgxml"
+                z.extract(name, outpath)
+                
+                fhr = open("../data/ipgxml/" + name, "r")
+                fhw = open("../data/ipgxml/" + name + ".tmp", "w")
 
-            fhr = open("../data/ipgxml/" + name, "r")
-            fhw = open("../data/ipgxml/" + name + ".tmp", "w")
+                fhw.write('<?xml version="1.0" encoding="UTF-8"?>\n')
+                fhw.write('<!DOCTYPE us-patent-grant SYSTEM "us-patent-grant-v45-2014-04-03.dtd" [ ]>\n')
+                fhw.write("<myroot>\n")
 
-            fhw.write('<?xml version="1.0" encoding="UTF-8"?>\n')
-            fhw.write('<!DOCTYPE us-patent-grant SYSTEM "us-patent-grant-v45-2014-04-03.dtd" [ ]>\n')
-            fhw.write("<myroot>\n")
-
-            for line in fhr:
-                if line.startswith("<?xml") or line.startswith("<!DOCTYPE"):
-                    continue
-                fhw.write(line)
+                for line in fhr:
+                    if line.startswith("<?xml") or line.startswith("<!DOCTYPE"):
+                        continue
+                    fhw.write(line)
             
-            fhw.write("</myroot>\n")
+                fhw.write("</myroot>\n")
 
-            fhr.close()
-            fhw.close()
-            os.remove("../data/ipgxml/" + name)
-            os.rename("../data/ipgxml/" + name + ".tmp", "../data/ipgxml/" + name)
+                fhr.close()
+                fhw.close()
+                os.remove("../data/ipgxml/" + name)
+                os.rename("../data/ipgxml/" + name + ".tmp", "../data/ipgxml/" + name)
             file_count += 1
             print("Processed " + str(file_count) + " files")
